Changed neighbors collection and data splitting method, added efficiency tests
We are now working with a max heap for the neighbors. We are able to use this collection because I save each neighbors' distance from the instance in it's "weight()" provided by WEKA, so we don't need a Tree Map for example. First I initialized the neighbors with the first k instances of the current (after split) training set, after that I go through all of the training instances and get the k closest neighbors. All other functions that rely on the neighbors collection also now work with the priority queue (heap). I also added the efficient distance functions and a time tracker for measuring the prediction time, and the relevant efficiency prints in Main.